requirements.txt
Flask==2.2.2
Flask-WTF==1.0.1
Werkzeug==2.2.2
tqdm==4.64.1
python-docx
openai
loguru
langchain 
langchain-openai 
langchain-experimental
faiss-cpu
unstructured
langchain[docarray]
numpy==1.26
Flask-SQLAlchemy==3.0.2
Flask-Login==0.6.2
Flask-Migrate==4.0.4
psycopg2-binary==2.9.6
python-dotenv==1.0.0
email_validator==2.2.0